web-app/requirements.txt
python-dotenv
Flask
boto3
faiss-cpu
langchain-aws
langchain-core
langchain-community
langchain-text-splitters
